v0.6.3.2-alpha stuff.

+++ CPPZombs/Project1/Framebuffers.h
@@ -109,7 +109,7 @@ public:
         // create position attachment texture
         glGenTextures(1, &positionBuffer);
         glBindTexture(GL_TEXTURE_2D, positionBuffer);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, width, height, 0, GL_RGB, GL_UNSIGNED_BYTE, NULL);
+        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, width, height, 0, GL_RGB, GL_UNSIGNED_BYTE, NULL);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
         glBlendEquation(GL_FUNC_ADD);
